Started working on the chat functions
Right now the chat box will display the interacted entity's chat
function, which is created when the creature is substantiated. Need to
add a function to the dialogue handler to recreate the chat available
when the player selects things, so as to allow for dialogue trees.
Should also consider adding a GOTO tag for the dialogue options so the
chat can have conversation trees.

Right now the player state changes from a moving to a chat state, so
input needs to be changed accordingly, allowing for the new state input.

All in all a pretty neat day.

Also, the XML is completely easy to use and expansive for the
playerbase, allowing modding. Noice.