Added chat boxes and a temporary selector for chat text
Summary says it all. Pretty much had to go back and reallocate the chat
drawing inside the main state, since it'll be the only one that ever
uses chat.

For the time, chat is a simple blue box with a pointer I jacked from the
net until I can find a better way to demonstrate information