CHAT ENGINE LIVVVVVVES!
A rudimentary selection/state machine is in operation for the chat
engine. It can parse a default xml at the moment, but it works!

Next I need a specific way to create dialogue to pop on the screen for
the NPC's response, allow wait for key press by player, then display the
new options the player can choose from.

Also, some form of scrolling for the chat available.

And word wrapping.

BUT IT LIVESSSSS.

Biggest issue was having the E key check in player.interact() not be
assigned solely to the movement state. Once I fixed that, everything
else worked as planned.