Postfix: Added "Goodbye", fixed chat redundancy
Added a goodbye to the bottom of all chats, so players can return to
moving when complete with conversation.

Can later add tag to scan xml for custom goodbyes from NPCS.

Also, fixed issue with chat redundancy. If a player has talked to an NPC
about an issue before, no need to talk about it again and take up
playerKnowledge[i][x].