Set up a response system for the dialogue handler
Now the XML has a response tag. If there isn't one, it goes back to the
selection process. If there is one, input defaults to a new method in
the DialogHandler called continueResponse(). It just keeps churning out
responses until it runs out of them, at which point, back to the
selection process.

I think I've written myself in circles due to not understanding the core
mechanics of the Element system, and I am quite certain all this could
have been done much easier and cleaner, however I am still proud of this
attempt and the fact that I stuck with it until it worked.

Next up, I could get the saidBy tag to print the name in one of the
corners with a picture in the opposite corner if one is available.

I should also work on word wrapping text. I think I have a solution in
using a simple check for current total char length compared to the new
char length after a space bar break, then adding the current chars to an
ArrayList<String> and repeating the process.

All in all, a good day.