Prototype for inventory system works
Basic setup, right now items picked up in the world will add it to a
grid system based on char stats. Each item in the itemlist.xml now has
its own width and height attributes available which translates in the
grid system.

Grid interaction with the cursor, as well as the cursor texture need to
be worked on.

All in all, neat day.