Inventory Manager updated; Implemented into trade manager
Now the inventory Manager will swap items that are held with items in
inv if there is room.

Trade manager now implements an inventory Manager for each part of the
trade. Player gets one, entity gets one, trade gets one(acting as the
barter medium). Player can't add item directly to entity inventory.
Items cannot be taken from entity inventory directly to player
inventory. For a trade to be established, items belonging to the player
must equal or exceed the value of the items requested.

TODO: Add some form of identification for trade items to show their
ownership(Player or entity)

Reinstate a total value indicator again to show the loss or profit of
the trade