Figured out why dog got a new inventory everytime he traded.
tl;dr I'm a moron.

In an attempt to fix the dog's speed loading on initialization, I set
his inventory manager to be built when a new trade manager was
established. Unfortunately, I didn't realize I had called this every
time a trade handler was established so the trade handler was showing a
brand new inventory.

Good news is I had a chance to pour over the trade handler and clean it
up some more. I think it should probably be swapped over to a component
for the main state to have, instead of something for the player to
establish