Got sick, added character Creation
Character creation is underway. Working on the gui and gameplay aspects.
Right now the new game takes you to the Character Creation State which
will feed you through a series of pages in the process, each containing
the main stats of the player. Done with first page, need to assign
buttons in the second page.